refactor: introduce structuring folders and fix namespaces
