+++ src/EVAutoreg/Interfaces/IMailEventListener.cs
@@ -1,7 +1,7 @@
 using Microsoft.Exchange.WebServices.Data;
 using Task = System.Threading.Tasks.Task;
 
-namespace EVAutoreg;
+namespace EVAutoreg.Interfaces;
 
 public interface IMailEventListener
 {
